Raw Data Type Logic revamped
